Adding pg

+++ db/migrate/20150212200935_create_users.rb
@@ -1,8 +1,8 @@
 class CreateUsers < ActiveRecord::Migration
   def change
     create_table :users do |t|
-		t.integer :user_id
-    	t.integer :room_id
+		  t.bigint :user_id
+    	t.bigint  :room_id
     	t.boolean :user_free
     	t.string  :user_name
     	t.string  :user_issue
